+++ features/step_definitions/packet_data_steps.rb
@@ -11,6 +11,14 @@ Given(/^a packet data file "(.*?)"$/) do |name|
   end
 end
 
+When(/^I create an OpenFlow message with "([^"]*)"$/) do |ruby_code|
+  @result = Pio.module_eval(ruby_code)
+end
+
+When(/^I create a packet with:$/) do |ruby_code|
+  @result = Pio.module_eval(ruby_code)
+end
+
 When(/^I try to parse the file with "(.*?)" class$/) do |parser|
   parser_klass = Pio.const_get(parser)
   if @raw
@@ -36,7 +44,16 @@ Then(/^it should finish successfully$/) do
   # Noop.
 end
 
-Then(/^the parsed data have the following field and value:$/) do |table|
+Then(/^the packet have the following field and value:$/) do |table|
+  table.hashes.each do |each|
+    output = each['field'].split('.').inject(@result) do |memo, method|
+      memo.__send__(method)
+    end
+    expect(output.to_s).to eq(each['value'])
+  end
+end
+
+Then(/^the message have the following field and value:$/) do |table|
   table.hashes.each do |each|
     output = each['field'].split('.').inject(@result) do |memo, method|
       memo.__send__(method)
@@ -46,7 +63,7 @@ Then(/^the parsed data have the following field and value:$/) do |table|
 end
 
 # rubocop:disable LineLength
-Then(/^the parsed data \#(\d+) have the following field and value:$/) do |index, table|
+Then(/^the message \#(\d+) have the following field and value:$/) do |index, table|
   table.hashes.each do |each|
     output = each['field'].split('.').inject(@result[index.to_i - 1]) do |memo, method|
       memo.__send__(method)